requirements.txt
asgiref==3.7.2
Django==5.0.1
sqlparse==0.4.4
tzdata==2023.4
djangorestframework==3.14.0
markdown==3.5.2
django-filter==23.5
drf-spectacular==0.27.0